# spec/api_contract.yaml
openapi: 3.0.3
info:
  title: ChronicleCore Local API
  description: |
    Local-only REST API for ChronicleCore time tracking application.
    All endpoints bind to 127.0.0.1 only. No remote access permitted.
  version: 1.0.0
  contact:
    name: ChronicleCore
servers:
  - url: http://127.0.0.1:{port}
    description: Local development server
    variables:
      port:
        default: '8080'
        description: Configurable port (default 8080)
...
paths:
...
  /api/v1/tracking/start:
    post:
      summary: Start tracking
      tags: [Tracking]
      description: Begin capturing activity (requires tracking to be stopped)
      responses:
        '200':
          description: Tracking started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackingStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/tracking/pause:
    post:
      summary: Pause tracking
      tags: [Tracking]
      description: Temporarily pause (requires active tracking)
      responses:
        '200':
          description: Tracking paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackingStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/tracking/resume:
    post:
      summary: Resume tracking
      tags: [Tracking]
      description: Resume from paused state
      responses:
        '200':
          description: Tracking resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackingStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/tracking/stop:
    post:
      summary: Stop tracking
      tags: [Tracking]
      description: Fully stop tracking (can restart later)
      responses:
        '200':
          description: Tracking stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackingStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/blocks:
    get:
      summary: List time blocks
      tags: [Blocks]
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter by date (YYYY-MM-DD). Defaults to today.
        - name: profile_id
          in: query
          required: false
          schema:
            type: integer
          description: Filter by profile
        - name: unassigned
          in: query
          required: false
          schema:
            type: boolean
          description: Show only unassigned blocks
      responses:
        '200':
          description: List of blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Block'
...
  /api/v1/profiles:
    get:
      summary: List profiles
      tags: [Profiles]
      parameters:
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Profile'

    post:
      summary: Create profile
      tags: [Profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileCreate'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/clients:
    get:
      summary: List clients
      tags: [Profiles]
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'

    post:
      summary: Create client
      tags: [Profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
  /api/v1/export/invoice-lines:
    post:
      summary: Export invoice CSV
      tags: [Exports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [start_date, end_date]
              properties:
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                profile_ids:
                  type: array
                  items:
                    type: integer
                  description: Filter by profiles (optional)
                rounding_minutes:
                  type: integer
                  enum: [6, 15]
                  default: 6
                  description: Rounding increment
                minimum_billable_minutes:
                  type: integer
                  default: 0
      responses:
        '200':
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
components:
...
  schemas:
    TrackingStatus:
      type: object
      properties:
        state:
          type: string
          enum: [STOPPED, ACTIVE, PAUSED]
        last_active_at:
          type: string
          format: date-time
          nullable: true
        idle_seconds:
          type: integer
        current_window:
          type: object
          nullable: true
          properties:
            app_name:
              type: string
            title:
              type: string
    Block:
      type: object
      properties:
        block_id:
          type: integer
        ts_start:
          type: string
          format: date-time
        ts_end:
          type: string
          format: date-time
        duration_minutes:
          type: number
          format: float
        primary_app_name:
          type: string
        primary_domain:
          type: string
          nullable: true
        title_summary:
          type: string
          nullable: true
        profile_id:
          type: integer
          nullable: true
        confidence:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        billable:
          type: boolean
        locked:
          type: boolean
        description:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    Profile:
      type: object
      properties:
        profile_id:
          type: integer
        client_name:
          type: string
        project_name:
          type: string
          nullable: true
        service_name:
          type: string
        rate_name:
          type: string
        rate_amount:
          type: number
          description: Hourly rate in major units
        currency:
          type: string
        is_active:
          type: boolean
    ProfileCreate:
      type: object
      required: [client_id, service_id, rate_id]
      properties:
        client_id:
          type: integer
        project_id:
          type: integer
          nullable: true
        service_id:
          type: integer
        rate_id:
          type: integer
        name:
          type: string
          nullable: true
    Client:
      type: object
      properties:
        client_id:
          type: integer
        name:
          type: string
        is_active:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - NOT_FOUND
                - INTERNAL_ERROR
                - INVALID_STATE
            message:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INVALID_REQUEST
              message: Invalid parameter value
              details: {}
...
security: []